fix: use f64 for volume to avoid f32 precision loss, improve config errors
f32 cannot represent 0.8 exactly, causing TOML serialisation to write
0.800000011920929 in the config file. Switching to f64 round-trips
cleanly as 0.8. Cast to f32 at the rodio call site.

Also improve the Config::load() error context to include the config file
path and a hint to delete the file to regenerate defaults, making bad
option errors (e.g. unknown position variant) easier to act on.

+++ src/daemon/sound.rs
@@ -21,7 +21,7 @@ impl SoundPlayer {
             return;
         }
         let path = crate::config::Config::expand_path(&self.config.file);
-        let volume = self.config.volume;
+        let volume = self.config.volume as f32;
         std::thread::spawn(move || {
             if let Err(e) = play_file(&path, volume) {
                 eprintln!("cooee: sound error: {}", e);
